+++ lib/ui/page/home/tab/chats/widget/search_user_tile.dart
@@ -19,14 +19,19 @@ import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
+import '/api/backend/schema.dart';
 import '/domain/model/chat.dart';
+import '/domain/model/precise_date_time/precise_date_time.dart';
 import '/domain/model/user.dart';
+import '/domain/model/welcome_message.dart';
 import '/domain/repository/contact.dart';
 import '/domain/repository/user.dart';
 import '/l10n/l10n.dart';
 import '/routes.dart';
 import '/themes.dart';
+import '/ui/page/home/page/user/controller.dart';
 import '/ui/page/home/widget/contact_tile.dart';
+import 'rectangle_attachment.dart';
 
 /// [ContactTile] intended to be used as a search result representing the
 /// provided [User] or [ChatContact].
@@ -63,6 +68,74 @@ class SearchUserTile extends StatelessWidget {
           ) !=
           null;
 
+      Widget? subtitle;
+
+      final WelcomeMessage? welcome = user?.user.value.welcomeMessage;
+
+      if (welcome != null) {
+        final List<Widget> images = [];
+        final String text = welcome.text?.val ?? '';
+
+        if (welcome.attachments.isNotEmpty) {
+          if (welcome.text == null) {
+            images.addAll(
+              welcome.attachments.map((e) {
+                return Padding(
+                  padding: const EdgeInsets.only(right: 2),
+                  child: RectangleAttachment(e),
+                );
+              }),
+            );
+          } else {
+            images.add(
+              Padding(
+                padding: const EdgeInsets.only(right: 4),
+                child: RectangleAttachment(welcome.attachments.first),
+              ),
+            );
+          }
+        }
+
+        subtitle = Row(
+          children: [
+            if (text.isEmpty)
+              Flexible(
+                child: LayoutBuilder(
+                  builder: (_, constraints) {
+                    return Row(
+                      children: images
+                          .take((constraints.maxWidth / (30 + 4)).floor())
+                          .toList(),
+                    );
+                  },
+                ),
+              )
+            else
+              ...images,
+            if (text.isNotEmpty) Flexible(child: Text(text)),
+          ],
+        );
+      } else {
+        final String text;
+
+        final bool isOnline = user?.user.value.online == true;
+        final bool isAway =
+            isOnline && user?.user.value.presence == UserPresence.away;
+        final PreciseDateTime? lastSeenAt = user?.user.value.lastSeenAt;
+
+        if (isAway) {
+          text = 'label_away'.l10n;
+        } else if (isOnline) {
+          text = 'label_online'.l10n;
+        } else if (lastSeenAt != null) {
+          text = lastSeenAt.val.toDifferenceAgo();
+        } else {
+          text = 'label_offline'.l10n;
+        }
+
+        subtitle = Text(text);
+      }
+
       return Padding(
         padding: const EdgeInsets.symmetric(horizontal: 10),
         child: ContactTile(
@@ -73,12 +146,11 @@ class SearchUserTile extends StatelessWidget {
           selected: selected,
           subtitle: [
             const SizedBox(height: 5),
-            Text(
-              user?.user.value.bio?.val ??
-                  '${'label_num'.l10n}${'colon_space'.l10n}${(contact?.user.value?.user.value.num ?? user?.user.value.num)}',
+            DefaultTextStyle(
               style: selected
                   ? style.fonts.small.regular.onPrimary
                   : style.fonts.small.regular.secondary,
+              child: subtitle,
             ),
           ],
           trailing: [